cleaning up comments.
