--- encoder/encoder.go
@@ -1,6 +1,0 @@
-package encoder
-
-type Encoder interface {
-	Encode(item interface{}) ([]byte, error)
-	Decode(data []byte, destination interface{}) error
-}
